copy argument directives to synthetic input fields When creating a GRT for field arguments, we generate a synthetic input object definition that that models the argument values, where each argument is modeled as a field on the synthetic input object. This affords some nice re-use of the GRT base classes that expect input object types.
We don't currently set any applied directives on these synthetic input field values. It would be nice if we did! This would allow simplifying how we identify fields that should be represented as GlobalID values.

This PR transfers any applied directives on the arguments over to the synthetic input field definitions. A future PR will use this to simplify InputLikeBase's behavior when wrapping/unwrapping values.
Github-Change-Id: 961465
GitOrigin-RevId: 6d5b06da5482a1b5688bf51f6c6855dc903df168