GRTConv This PR creates a `GRTConv` object, which is a factory for `Conv`s that map between GRT values and their IR counterparts.
The nature of GRTs is that they are a thin type-safe wrapper around a tree of untyped engine data. This makes `GRTConv` fairly straightforward to implement as a thin layer of Convs that wrap/unwrap GRT classes and delegate everything else to `EngineValueConv`.

I'm posting this PR with some key tests disabled due to an issue with how default input values are managed:
default input values are written into an input objects backing map, which breaks roundtrippability for inputs. I plan on addressing this in a follow-up PR.
Github-Change-Id: 962091
GitOrigin-RevId: 79bf05ddbe7f9676a2be44acf4420ff61d952b0d