Add collectVariableDefinitions and fix collectVariableUsages This PR adds a function `collectVariableDefinitions`, which will be used in a follow-up PR that coerces resolved variables. This traverses a node (e.g. document, selection set) and finds all variable references, then constructs a single VariableDefinition for each variable name.
There is an existing similar function called `collectVariableUsages` that traverses a node and finds all variables of a particular name, constructing a `VariableUsageInfo` for each location. However, the traversal logic was incomplete and only handled very specific use cases, failing to find many variables. This PR fixes that logic and factors it out into a shared `collectAllVariableUsages` function.
Github-Change-Id: 961155
GitOrigin-RevId: 0a4c4c94f6aea5d223ef4db2ca3f6b7015b3a233